Rev 0.1 Carrier Board ready for Proto-Run
The Rev 0.1 Carrier board is just about ready for a prototype run on the
PCB mill and is designed for a mill with no silk or solder mask. Changes
include an edited schematic to include a better ESP part. power
considerations, and some formatting for readability. A first-run
mill-able board has been prepared as well. Finally, the Git Ignore was
modified to ignore temporary Eagle Files.